requirements.txt
numpy
matplotlib

torch
stheno

python-slugify
